VideoQnA-Demo/app/backend/requirements.txt
azure-identity==1.15.0
azure-search-documents==11.6.0b1
Flask==3.0.3
Flask-Limiter==3.6.0
openai==1.27.0
numpy==1.26.4
isodate==0.6.1
chromadb==0.5.0
tiktoken==0.7.0
